Move required Streaming API to sql/api
