+++ sql/api/src/main/scala/org/apache/spark/sql/catalyst/streaming/InternalOutputModes.scala
